requirements.txt
# Core ML/DL frameworks
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0

# Model architectures
timm>=0.9.0
efficientnet-pytorch>=0.7.1

# Experiment tracking and visualization
wandb>=0.15.0
tensorboard>=2.13.0

# Web interface
streamlit>=1.25.0

# Configuration management
pyyaml>=6.0

# Utilities
pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0
